introduce isl_args structure that constains isl_arg list and size of options
Before, the size of the options structure of a submodule was stored
inside the corresponding child option of the parent.
This however, requires that the options structure of the submodule
is visible to the parent so that it can compute the size of this structure.
Moreover, if the size of this structure changes, then the code
containing the child option, which may reside in a different module,
would have to be recompiled.

Now, we store the size of the options structure together with the list
of option descriptions.  In particular, it is stored in the same module
as the options structure itself, so that this structure no longer needs
to be exposed.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>