isl_arg_parse: print current option values as defaults
In the --help output, we used to print the default values stored
in the isl_arg objects as the defaults for options.
It is however sometimes convenient to change some of the options
to values different from their true defaults before allowing
the user to override these options.  It then makes more sense
to print the current values of the options as defaults, rather
than the defaults that were used during the construction of
options object.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>