isl_aff_nonneg_basic_set and isl_aff_zero_basic_set: simplify result
The constructed constraint may be a tautology in which case it needs
to be simplified away, especially if it is an equality.

We probably want to perform this simplification in
isl_basic_set_from_constraint instead, but CLooG currently depends
on the resulting basic set not being simplified.
We would have to fix CLooG first before we can change this behavior.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>