schedule.c: split_parallel: avoid invalid memory accesses
The original code incorrectly assumed that the schedules for
all statements have the same number of columns, i.e., that the
dimensions of the iteration domains of all statements are the same.

We fix this up by simply bailing out if we find this assumption not to hold.
We will improve the applicability of the transformation performed by
split_parallel in a subsequent patch.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>