isl_basic_set_full_compression: detect equalities in input first
Without this explicit detection of equalities, some equalities
may be discovered in the middle of the computation, possibly
leading to the introduction of existentially quantified variables,
while callers of isl_basic_set_full_compression typically do not
want such variables.

Since the detection of equalities can in some cases be fairly expensive,
we may have to find a better way of dealing with this problem at some point.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>