isl_schedule_get_map: extract schedule from band forest if available
In the next commits, we will allow the user to modify the band forest.
This means that the schedule represented by the modified band forest
may not be the same as the schedule that was originally computed.
We therefore need to extract it from this forest if the forest
has been constructed already.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>