isl_access_info: change interface for specifying restrictions
The original interface would only allow the user to restrict the source
iterations, while the user may also want to restrict the sink iterations.
The new interface also makes it more convenient for the user to specify
that no restriction should be applied and allows for a restriction
on the result of the maximization problem.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>