isl_union_set_compute_schedule: also keep track of proximity edges in hash table
We don't use this additional hash table yet, but we will need it the next
commit. A separate hash table is needed because the entries are keyed
by the source and destination node, so there can be only one edge
between a given pair of nodes in a single table.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>