isl_multi_*_reset_space_and_domain: call isl_*_reset_domain_space
The original code would call isl_*_reset_space_and_domain,
but the space of the multi_* is clearly different from the space of the *,
so it does not make sense to pass that space to the *.
In case of isl_multi_aff, this does not cause any problems because
isl_aff_reset_space_and_domain ignore the first argument.
However, for an isl_multi_pw_aff, this would cause problems.
Call isl_*_reset_domain_space instead.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>