isl_basic_map_drop_redundant_divs: don't drop divs that appear in other divs
Divs that appear in the definition of other divs usually occur in at least
4 constraints, so they normally wouldn't get dropped by this function.
However, the div constraints of the other div(s) may have been simplified
or eliminated.  This happens in particular on the new test case
due to the optimization introduced in 2592280 (isl_basic_map_simplify:
eliminate known divs that appear with unit coefficient,
Wed May 16 13:51:16 2012 +0200).

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>