isl_ast_codegen.c: generate_domain: postpone single-valuedness test
In particular, perform the test on the map that will actually be used
to construct the call expression.

The original code would first perform the test and then construct
the call expression from a map with a simplified domain.
This construction may fail because after simplification of the domain,
the map may no longer be single valued.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>