isl_schedule.c: extract_edge: look for matching edge directly
If there are two edges of different types that happen to have the
same map field, then we want to merge the two edges into one.
Before, we would look for any edge (of a different type) between
the same to nodes and then check if the map fields happen to be the same.
Now we continue looking for edges of other types if the map fields are
not the same.

This change should not have any effect in the current code base
because there are only two types of edges, but it will have an
effect once we consider more types.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>