isl_schedule_get_band_forest: sort bands in band list
In particular if the partial schedule of the parent of a band list
assigns different constant values to the domains of the bands in the list,
then sort the bands according to those constant values.

This should provide for a more "natural" order to the user.
The need for sorting could be avoided entirely if we were to
construct a forest of bands directly from isl_union_set_compute_schedule.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>