isl_basic_map_compute_divs: preserve already known divs
The purpose of isl_basic_map_compute_divs is to ensure that all divs
have an explicit representation.  If we already have an explicit
representation for some of the divs, then there is no need to
recompute an explicit representation for them.  Instead, we preserve
the explicit representations of the known divs and only compute
explicit representations for the currently unknown divs.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>