isl_basic_set_parameter_compression: allow divs in input
For the purpose of compressing parameters, there should be no need
to make a distinction between set variables and existentially quantified
variables.
There does not appear to be any reason for not allowing existentially
quantified variables before other than the fact that we didn't need
to apply this function of basic sets with existentially quantified
variables before.  However, this will turn out to be useful when
we want to compress the unknown divs within isl_basic_map_compute_divs.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>