isl_basic_map_lexopt: preinitialize domain
The original code would pass a universe domain to isl_basic_map_partial_lexopt,
but this means that the parametric integer programming engine will consider
all inequalities as splitting (with respect to the universe context),
resulting in a fair amount of wasted effort.
More importantly, the border constraints may be transferred to the
context in the middle of constraints that actually split the context
into parts with possibly different solutions.  The current heuristic
for merging solutions that turn out to be identical will not be able
to transfer solution across the border constraints since they only
allow for a solution on one side of the constraint.

By prepopulating the domain with some of the constraints from the input map,
we can avoid some redundant computations and we improve the opportunities
for merging identical partial solutions.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>