isl_basic_map_compute_divs: recursively remove parameter equalities
The removal of explicit equalities among the parameters may uncover
additional equalities.  If so, we now recurse in parameter_compute_divs
so that also these extra equalities are removed.
This will be important for the next commit where we will try to compress
the existentially quantified variables and this requires the absence
of equalities among the parameters.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>