isl_tab_basic_map_partial_lexopt: disable use of shifted tableau on equalities
If the input context already involves equalities then the shfited tableau
is not used as it would be empty.  However, if equalities are added at
a later stage (e.g., transferred for the input map), then we would fail
to disable the use of the shifted tableau, possibly resulting in invalid
sample values being computing from this shifted tableau.

Reported-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>