isl_schedule_get_map: pad schedule map with zeros if reconstructed from bands
The schedule map extracted from the isl_schedule_nodes is padded with zeros
by construction.  However, if the schedule map is reconstructed from the
band forest, then the schedule map was not be padded with zeros.
Pad this schedule as well.

We may want to change isl_band_list_get_suffix_schedule to perform this
padded instead.

Reported-by: Ali El Moussawi <elmoussawi.ali@gmail.com>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>