isl_ast_codegen.c: create_node_scaled: reflect current domain slice in schedule
If we detect that a given schedule dimension has a fixed (possibly
parametric) value, then this value is recorded in the build and
the schedule dimension is eliminated.  This elimination process may
drop constraints on the part of the domain that we are currently considering,
which may cause all sorts of problems.  In particular, we may fail
to discover at the end of the AST generation that the original schedule
was single valued and insert spurious loops.

Plug in the value we found instead of simply eliminating the schedule
dimension.

Reported-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>