isl_ast_build_get_stride_constraint: combine multiple strides
If we find multiple constraints from which can derive a stride,
then we would only keep the stride information from the last
of these constraints.  We now combine multiple strides into
a single stride.  Note that this may not have any effect in practice
because multiple strides are usually already combined into one
at the level of the constraints of a basic set.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>